+++ resources/views/livewire/front/passager/passager-data-table.blade.php
@@ -3,9 +3,11 @@
     <x-front.card>
         <x-front.title>
             {{ __('Liste des passagers') }}
-            <x-slot:button>
-                <x-button primary label="{{ __('Créer un passager') }}" icon="plus" href="{{ route('front.passager.create') }}"/>
-            </x-slot:button>
+            @can('create passenger')
+                <x-slot:button>
+                    <x-button primary label="{{ __('Créer un passager') }}" icon="plus" href="{{ route('front.passager.create') }}"/>
+                </x-slot:button>
+            @endcan
         </x-front.title>
 
         <x-datatable>
@@ -30,8 +32,12 @@
                         </x-datatable.td>
                         <x-datatable.td>
                             <div class="space-x-2">
-                                <x-button.circle icon="pencil" primary href="{{ route('front.passager.edit', ['passager' => $passager->id]) }}"/>
-                                <x-button.circle icon="trash" red wire:click="deletePassenger({{ $passager }})"/>
+                                @can('edit passenger')
+                                    <x-button.circle icon="pencil" primary href="{{ route('front.passager.edit', ['passager' => $passager->id]) }}"/>
+                                @endcan
+                                @can('delete passenger')
+                                    <x-button.circle icon="trash" red wire:click="deletePassenger({{ $passager }})"/>
+                                @endcan
                             </div>
                         </x-datatable.td>
                     </x-datatable.tr>
